+++ extension.js
@@ -826,6 +826,8 @@ function activate(context) {
 	follow_cursor_status_bar.tooltip =
 		"When enabled, keep editor cursor and Ren'Py in sync"
 
+	let skip_next_current_line = false
+
 	pm = new ProcessManager({
 		stdout_callback({ data, game_root, is_supports_exec_py }) {
 			if (data.startsWith('RENPY_WARP_SIGNAL_CURRENT_LINE')) {
@@ -833,6 +835,12 @@ function activate(context) {
 				const [, abs_path, local_path, line] = data.trim().split(':')
 
 				logger.debug(`renpy reports line: ${local_path}:${line}`)
+
+				if (skip_next_current_line) {
+					logger.info('skipped redundant current line sync')
+					skip_next_current_line = false
+					return
+				}
 
 				return sync_editor_with_renpy({
 					abs_path,
@@ -850,6 +858,7 @@ function activate(context) {
 				logger.info(
 					'game reload detected. attempting to re-inject sync script'
 				)
+				skip_next_current_line = true
 				return inject_sync_script(game_root)
 			}
 		},